Remove task yield and make blocking APIs non-blocking
- Task manager: remove pifTaskManager_Yield*/YieldAbort* and the
  disallow_yield_id / _running fields. A loop callback now always runs
  to its end; a task that needs to wait returns and is released again
  by its delay or pifTask_SetTrigger().
- pif: pif_Delay1ms/1us now take uint32_t and are documented as
  CPU-holding waits. Drivers that used Yield (i2c scan, pmlcd_i2c)
  now call these instead.
- Log: pifLog_PrintInBuffer() sends only what fits and returns TRUE
  while data remains. Add pifLog_DropCount()/ResetDropCount().
- GPS u-blox: remove the blocking parameter from the request functions
  and add pifGpsUblox_CheckRequest() to poll the result.
- Modbus RTU/ASCII master: requests return once the frame is queued.
  Add pifModbus*Master_Check() and PifModbusMasterResult.
- Touch screen: replace pifTouchScreen_Calibration() with
  pifTouchScreen_StartCalibration(), a state machine driven by the
  touch task that reports the result through a callback.
- BASIC interpreter: run programs step by step across task releases
  (PifBasicState) instead of in a single call.
- ADS1x1x: replace ReadMux/VoltageMux with pifAds1x1x_StartMux() and
  timer-based conversion (pifAds1x1x_AttachTimer); short conversions
  still spin.
- MS5611: replace ReadRaw*/ReadBarometric with timer-driven
  pifMs5611_StartBarometric() and pifMs5611_AttachTimer().
- BMP280: add pifBmp280_StartMeasurement(); ReadRawData/ReadBarometric
  no longer wait and return FALSE until the sample is ready.
- Storage (base/fix/var): Read/Write return PifStorageStart. Multi-sector
  transfers are carried by a timer (pifStorage_AttachTimer), and the
  write cycle is tracked instead of waited out after each write.
- ILI9341/SSD1963: fill a rectangle with one address window rather than
  line by line.
- Update comments and config templates to match.

+++ include/core/pif.h
@@ -210,17 +210,21 @@ void pif_sigTimer1ms();
 
 /**
  * @fn pif_Delay1ms
- * @brief Performs a delay operation in the core context for the requested time interval.
- * @param delay Delay duration value.
+ * @brief Waits for the requested time with the CPU held. Nothing else runs meanwhile, so the whole
+ *        wait is added to the block time of whichever task is executing and delays a TM_REALTIME
+ *        release by that much. Keep it to initialization and to waits a device cannot be asked
+ *        about; a wait a task can be released from again belongs in the return value of its loop
+ *        callback instead.
+ * @param delay Delay duration in milliseconds.
  */
-void pif_Delay1ms(uint16_t delay);
+void pif_Delay1ms(uint32_t delay);
 
 /**
  * @fn pif_Delay1us
- * @brief Performs a delay operation in the core context for the requested time interval.
- * @param delay Delay duration value.
+ * @brief Waits for the requested time with the CPU held, as pif_Delay1ms() does.
+ * @param delay Delay duration in microseconds.
  */
-void pif_Delay1us(uint16_t delay);
+void pif_Delay1us(uint32_t delay);
 
 /**
  * @fn pif_ChangeStatusLed
